spliting into `kindelia_core` and `kindelia`
`kindelia_core`: core node logic
`kindelia`: CLI, HTTP server, HTTP client
