+++ kindelia/src/main.rs
@@ -1,9 +1,633 @@
-#![allow(clippy::single_component_path_imports)]
+mod api;
 mod cli;
+mod config;
+mod files;
+mod util;
 
-pub use clap::{Parser, Subcommand};
+use std::future::Future;
+use std::net::UdpSocket;
+
+use api::client::ApiClient;
+use clap::Parser;
+
+use cli::{
+  Cli, CliCommand, GetCtrKind, GetFunKind, GetKind, GetRegKind, NodeCommand,
+  UtilCommand,
+};
+use config::{arg_from_file_or_stdin, ConfigSettingsBuilder};
+
+use files::FileInput;
+use kindelia_core::api::{Hash, HexStatement};
+use kindelia_core::bits::ProtoSerialize;
+use kindelia_core::common::Name;
+use kindelia_core::config::{ApiConfig, MineConfig, NodeConfig, UiConfig};
+use kindelia_core::crypto;
+use kindelia_core::events;
+use kindelia_core::hvm::{self, view_statement, Statement};
+use kindelia_core::net;
+use kindelia_core::net::ProtoComm;
+use kindelia_core::node;
+use kindelia_core::node::{spawn_miner, Node};
+use kindelia_core::util::bytes_to_bitvec;
+use util::{
+  bytes_to_u128, flag_to_option, handle_config_file, run_async_blocking,
+};
+
+use crate::cli::GetStatsKind;
+use crate::util::init_config_file;
 
 fn main() -> Result<(), String> {
-  cli::run_cli()?;
+  run_cli()
+}
+
+// Macros
+// ======
+
+macro_rules! resolve_cfg {
+  // Resolve config value with no file
+  // ---------------------------------
+
+  // With default value
+  (env = $env:expr, default_fn = $default:expr, cli_val = $cli:expr $(,)*) => {
+    ConfigSettingsBuilder::default()
+      .env($env)
+      .default_value($default)
+      .build()
+      .unwrap()
+      .resolve($cli, None)?
+  };
+
+  // No default value
+  (env = $env:expr, no_default_fn = $default:expr, cli_val = $cli:expr $(,)*) => {
+    ConfigSettingsBuilder::default()
+      .env($env)
+      .prop($prop)
+      .default_value($default)
+      .build()
+      .unwrap()
+      .resolve($cli, None)?
+  };
+
+  // Resolve config value with file
+  // ------------------------------
+
+  // With default value
+  (env = $env:expr, prop = $prop:expr, default = $default:expr, cli_val = $cli:expr, cfg = $cfg:expr $(,)*) => {
+    ConfigSettingsBuilder::default()
+      .env($env)
+      .prop($prop)
+      .default_value(|| Ok($default))
+      .build()
+      .unwrap()
+      .resolve($cli, $cfg)?
+  };
+
+  // No default value
+  (env = $env:expr, prop = $prop:expr, no_default = $default:expr, cli_val = $cli:expr, cfg = $cfg:expr $(,)*) => {
+    ConfigSettingsBuilder::default()
+      .env($env)
+      .prop($prop)
+      .default_value(|| Err($default))
+      .build()
+      .unwrap()
+      .resolve($cli, $cfg)?
+  };
+}
+
+/// Parse CLI arguments and run
+pub fn run_cli() -> Result<(), String> {
+  let parsed = Cli::parse();
+  let default_base_path = || {
+    let home_dir = dirs::home_dir().ok_or("Could not find $HOME path")?;
+    Ok::<_, String>(home_dir.join(".kindelia"))
+  };
+  let default_node_data_path =
+    || Ok::<_, String>(default_base_path()?.join("state"));
+  let default_config_path = || Ok(default_base_path()?.join("kindelia.toml"));
+  let default_api_url = || Ok("http://localhost:8000".to_string());
+
+  // Resolve path where to load config file from
+  let config_path = resolve_cfg!(
+    env = "KINDELIA_CONFIG",
+    default_fn = default_config_path,
+    cli_val = parsed.config,
+  );
+
+  let api_url = resolve_cfg!(
+    env = "KINDELIA_API_URL",
+    default_fn = default_api_url,
+    cli_val = parsed.api,
+  );
+
+  match parsed.command {
+    CliCommand::Test { file, sudo } => {
+      let code: String = file.read_to_string()?;
+      test_code(&code, sudo);
+      Ok(())
+    }
+    CliCommand::Serialize { file } => {
+      let code: String = file.read_to_string()?;
+      serialize_code(&code);
+      Ok(())
+    }
+    CliCommand::Deserialize { file } => {
+      let code: String = file.read_to_string()?;
+      deserialize_code(&code)
+    }
+    CliCommand::Unserialize { stmt } => deserialize_code(&stmt),
+    CliCommand::Sign { file, secret_file, encoded, encoded_output } => {
+      let skey: String = arg_from_file_or_stdin(secret_file.into())?;
+      let skey = skey.trim();
+      let skey = hex::decode(skey).map_err(|err| {
+        format!("Secret key should be valid hex string: {}", err)
+      })?;
+      let skey: [u8; 32] = skey
+        .try_into()
+        .map_err(|_| "Secret key should have exactly 64 bytes".to_string())?;
+      let code = load_code(file, encoded)?;
+      let statement = match &code[..] {
+        [stmt] => sign_code(stmt, &skey),
+        _ => Err("Input file should contain exactly one statement".to_string()),
+      }?;
+      if encoded_output {
+        println!("{}", hex::encode(statement.proto_serialized().to_bytes()));
+      } else {
+        println!("{}", view_statement(&statement));
+      };
+      Ok(())
+    }
+    CliCommand::RunRemote { file, encoded } => {
+      // TODO: client timeout
+      let code = file.read_to_string()?;
+      let f = |client: api::client::ApiClient, stmts| async move {
+        client.run_code(stmts).await
+      };
+      let stmts = if encoded {
+        statements_from_hex_seq(&code)?
+      } else {
+        hvm::parse_code(&code)?
+      };
+      let results = run_on_remote(&api_url, stmts, f)?;
+      for result in results {
+        println!("{}", result);
+      }
+      Ok(())
+    }
+    CliCommand::Publish { file, encoded } => {
+      let code = file.read_to_string()?;
+      let stmts = if encoded {
+        statements_from_hex_seq(&code)?
+      } else {
+        hvm::parse_code(&code)?
+      };
+      publish_code(&api_url, stmts)
+    }
+    CliCommand::Post { stmt } => {
+      let stmts = statements_from_hex_seq(&stmt)?;
+      publish_code(&api_url, stmts)
+    }
+    CliCommand::Get { kind, json } => {
+      let prom = get_info(kind, json, &api_url);
+      run_async_blocking(prom)
+    }
+    CliCommand::Init => {
+      let path = default_config_path()?;
+      eprintln!("Writing default configuration to '{}'...", path.display());
+      init_config_file(&path)?;
+      Ok(())
+    }
+    CliCommand::Node { command, data_dir } => {
+      let config = handle_config_file(&config_path)?;
+      let config = Some(&config);
+
+      let data_path = resolve_cfg!(
+        env = "KINDELIA_NODE_DATA_DIR",
+        prop = "node.data.dir",
+        default = default_node_data_path()?,
+        cli_val = data_dir,
+        cfg = config,
+      );
+
+      match command {
+        NodeCommand::Clean => {
+          // warning
+          println!(
+            "WARNING! This will delete all the files present in '{}'...",
+            data_path.display()
+          );
+          // confirmation
+          println!("Do you want to continue? ['y' for YES / or else NO]");
+          let mut answer = String::new();
+          std::io::stdin()
+            .read_line(&mut answer)
+            .map_err(|err| format!("Could not read your answer: '{}'", err))?;
+          // only accept 'y' as positive answer, anything else will be ignored
+          if answer.trim().to_lowercase() == "y" {
+            std::fs::remove_dir_all(data_path).map_err(|err| {
+              format!("Could not remove the files: '{}'", err)
+            })?;
+            println!("All items were removed.");
+          } else {
+            println!("Canceling operation.");
+          }
+          Ok(())
+        }
+        NodeCommand::Start { initial_peers, network_id, mine, json } => {
+          // TODO: refactor config resolution out of command handling (how?)
+
+          // Get arguments from cli, env or config
+
+          let network_id = resolve_cfg!(
+            env = "KINDELIA_NETWORK_ID",
+            prop = "node.network.network_id",
+            no_default = "Missing `network_id` paramenter.".to_string(),
+            cli_val = network_id,
+            cfg = config,
+          );
+
+          let initial_peers = resolve_cfg!(
+            env = "KINDELIA_NODE_INITIAL_PEERS",
+            prop = "node.network.initial_peers",
+            default = vec![],
+            cli_val = initial_peers,
+            cfg = config,
+          );
+
+          let mine = resolve_cfg!(
+            env = "KINDELIA_MINE",
+            prop = "node.mining.enable",
+            default = false,
+            cli_val = flag_to_option(mine),
+            cfg = config,
+          );
+
+          let slow_mining = ConfigSettingsBuilder::default()
+            .env("KINDELIA_SLOW_MINING")
+            .prop("node.debug.slow_mining")
+            .default_value(|| Ok(0))
+            .build()
+            .unwrap()
+            .resolve_from_file_opt(config)?;
+
+          let api_config = ConfigSettingsBuilder::default()
+            .prop("node.api")
+            .default_value(|| Ok(ApiConfig::default()))
+            .build()
+            .unwrap()
+            .resolve_from_file_only(config)?;
+
+          // Start
+          let node_comm = init_socket().expect("Could not open a UDP socket");
+          let initial_peers = initial_peers
+            .iter()
+            .map(|x| net::parse_address(x))
+            .collect::<Vec<_>>();
+
+          let node_cfg = NodeConfig {
+            network_id,
+            data_path,
+            mining: MineConfig { enabled: mine, slow_mining },
+            ui: Some(UiConfig {
+              json,
+              tags: vec![events::NodeEventDiscriminant::Heartbeat],
+            }),
+            ws: None, // TODO: load from config file
+          };
+
+          let api_config = Some(api_config);
+          start_node(node_cfg, api_config, node_comm, initial_peers);
+
+          Ok(())
+        }
+      }
+    }
+    CliCommand::Util { command } => match command {
+      UtilCommand::DecodeName { file } => {
+        let txt = file.read_to_string()?;
+        let data: Result<Vec<Vec<u8>>, _> = txt
+          .trim()
+          .split(|c: char| c.is_whitespace())
+          .map(hex::decode)
+          .collect();
+        let data =
+          data.map_err(|err| format!("Invalid hex string: {}", err))?;
+        let nums = data.iter().map(|v| bytes_to_u128(v));
+        for num in nums {
+          if let Some(num) = num {
+            if let Ok(name) = Name::try_from(num) {
+              println!("{}", name);
+              continue;
+            }
+          }
+          println!();
+        }
+        Ok(())
+      }
+    },
+    CliCommand::Completion { .. } => todo!(),
+  }
+}
+
+// Client util
+// ===========
+
+fn run_on_remote<T, P, F>(
+  api_url: &str,
+  stmts: Vec<Statement>,
+  f: F,
+) -> Result<T, String>
+where
+  F: FnOnce(api::client::ApiClient, Vec<HexStatement>) -> P,
+  P: Future<Output = Result<T, String>>,
+{
+  let stmts: Vec<HexStatement> = stmts.into_iter().map(|s| s.into()).collect();
+  let client =
+    api::client::ApiClient::new(api_url, None).map_err(|e| e.to_string())?;
+  run_async_blocking(f(client, stmts))
+}
+
+// Client
+// ======
+
+/// This client is meant to talk with a node implementing UDP protocol
+/// communication (the default).
+type NC = UdpSocket;
+
+pub async fn get_info(
+  kind: GetKind,
+  json: bool,
+  host_url: &str,
+) -> Result<(), String> {
+  let client = ApiClient::new(host_url, None).map_err(|e| e.to_string())?;
+  match kind {
+    GetKind::BlockHash { index } => {
+      let block_hash = client.get_block_hash(index).await?;
+      println!("{}", block_hash);
+      Ok(())
+    }
+    GetKind::Block { hash } => {
+      let hash = Hash::try_from(hash.as_str())?;
+      let block = client.get_block(hash).await?;
+      println!("{:#?}", block);
+      Ok(())
+    }
+    GetKind::Ctr { name, stat } => {
+      let ctr_info = client.get_constructor(name).await?;
+      match stat {
+        GetCtrKind::Arity => {
+          println!("{}", ctr_info.arit)
+        }
+        GetCtrKind::Code => {
+          let args = (0..ctr_info.arit)
+            .map(|x| format!("x{}", x))
+            .collect::<Vec<_>>()
+            .join(" ");
+          println!("{{{} {}}}", name, args)
+        }
+      }
+      Ok(())
+    }
+    GetKind::Fun { name, stat } => match stat {
+      GetFunKind::Code => {
+        let func_info = client.get_function(name).await?;
+        if json {
+          println!("{}", serde_json::to_string(&func_info).unwrap());
+        } else {
+          let func = func_info.func;
+          let statement = hvm::Statement::Fun {
+            name,
+            args: vec![Name::NONE],
+            func,
+            init: Some(hvm::Term::var(Name::NONE)), // to show that we are actually not returning the initial state
+            sign: None,
+          };
+          println!("{}", statement);
+        }
+        Ok(())
+      }
+      GetFunKind::State => {
+        let state = client.get_function_state(name).await?;
+        if json {
+          println!("{}", serde_json::to_string_pretty(&state).unwrap());
+        } else {
+          println!("{}", state);
+        }
+        Ok(())
+      }
+      GetFunKind::Slots => todo!(),
+    },
+    GetKind::Reg { name, stat } => {
+      let reg_info = client.get_reg_info(&name).await?;
+      match stat {
+        GetRegKind::Owner => {
+          println!("{:x}", *(reg_info.ownr))
+        }
+        GetRegKind::List => {
+          for name in reg_info.stmt {
+            println!("{}", name)
+          }
+        }
+      }
+      Ok(())
+    }
+    GetKind::Stats { stat_kind } => {
+      let stats = client.get_stats().await?;
+      match stat_kind {
+        None => {
+          if json {
+            println!("{}", serde_json::to_string_pretty(&stats).unwrap());
+          } else {
+            println!("{:#?}", stats);
+          }
+        }
+        Some(stat_kind) => {
+          let val = match stat_kind {
+            GetStatsKind::Tick => stats.tick,
+            GetStatsKind::Mana => stats.mana,
+            GetStatsKind::Space => stats.space,
+            GetStatsKind::FunCount => stats.fun_count,
+            GetStatsKind::CtrCount => stats.ctr_count,
+            GetStatsKind::RegCount => stats.reg_count,
+          };
+          println!("{}", val);
+        }
+      };
+      Ok(())
+    }
+    GetKind::Peers { all } => {
+      let peers = client.get_peers::<NC>(all).await?;
+      for peer in peers {
+        println!("{}", peer.address)
+      }
+      Ok(())
+    }
+  }
+}
+
+// Code
+// ====
+
+pub fn serialize_code(code: &str) {
+  let statements =
+    hvm::read_statements(code).map_err(|err| err.erro).unwrap().1;
+  for statement in statements {
+    println!("{}", hex::encode(statement.proto_serialized().to_bytes()));
+  }
+}
+
+pub fn deserialize_code(content: &str) -> Result<(), String> {
+  let statements = statements_from_hex_seq(content)?;
+  for statement in statements {
+    println!("{}", view_statement(&statement))
+  }
   Ok(())
+}
+
+pub fn sign_code(
+  statement: &Statement,
+  skey: &[u8; 32],
+) -> Result<Statement, String> {
+  let user = crypto::Account::from_private_key(skey);
+  let hash = hvm::hash_statement(statement);
+  let sign = user.sign(&hash);
+  match statement {
+    Statement::Fun { sign, .. }
+    | Statement::Ctr { sign, .. }
+    | Statement::Run { sign, .. }
+    | Statement::Reg { sign, .. } => {
+      if sign.is_some() {
+        return Err("Statement already has a signature.".to_string());
+      }
+    }
+  };
+  let stat = hvm::set_sign(statement, sign);
+  Ok(stat)
+}
+
+fn load_code(file: FileInput, encoded: bool) -> Result<Vec<Statement>, String> {
+  let code = file.read_to_string()?;
+  handle_code(&code, encoded)
+}
+
+fn handle_code(code: &str, encoded: bool) -> Result<Vec<Statement>, String> {
+  if encoded {
+    statements_from_hex_seq(code)
+  } else {
+    hvm::parse_code(code)
+  }
+}
+
+fn statements_from_hex_seq(txt: &str) -> Result<Vec<Statement>, String> {
+  txt
+    .trim()
+    .split(|c: char| c.is_whitespace())
+    .map(statement_from_hex)
+    .collect()
+}
+
+fn statement_from_hex(hex: &str) -> Result<Statement, String> {
+  let bytes = hex::decode(hex)
+    .map_err(|err| format!("Invalid hexadecimal '{}': {}", hex, err))?;
+  hvm::Statement::proto_deserialized(&bytes_to_bitvec(&bytes))
+    .ok_or(format!("Failed to deserialize '{}'", hex))
+}
+pub fn publish_code(
+  api_url: &str,
+  stmts: Vec<Statement>,
+) -> Result<(), String> {
+  let f =
+    |client: ApiClient, stmts| async move { client.publish_code(stmts).await };
+  let results = run_on_remote(api_url, stmts, f)?;
+  for (i, result) in results.iter().enumerate() {
+    print!("Transaction #{}: ", i);
+    match result {
+      Ok(_) => println!("PUBLISHED (tx added to mempool)"),
+      Err(_) => {
+        println!("NOT PUBLISHED (tx is probably already on mempool)")
+      }
+    }
+  }
+  Ok(())
+}
+
+pub fn test_code(code: &str, sudo: bool) {
+  hvm::test_statements_from_code(code, sudo);
+}
+
+fn init_socket() -> Option<UdpSocket> {
+  let try_ports =
+    [net::UDP_PORT, net::UDP_PORT + 1, net::UDP_PORT + 2, net::UDP_PORT + 3];
+  for port in try_ports {
+    if let Ok(socket) = UdpSocket::bind(&format!("0.0.0.0:{}", port)) {
+      socket.set_nonblocking(true).ok();
+      return Some(socket);
+    }
+  }
+  None
+}
+
+// Node main thread
+// ================
+
+// TODO: I don't know why 'static is needed here or why it works
+pub fn start_node<C: ProtoComm + 'static>(
+  node_config: NodeConfig,
+  api_config: Option<ApiConfig>,
+  comm: C,
+  initial_peers: Vec<C::Address>,
+) {
+  eprintln!("Starting Kindelia node...");
+  eprintln!("Store path: {:?}", node_config.data_path);
+  eprintln!("Network ID: {:#X}", node_config.network_id);
+
+  // Threads
+  let mut threads = vec![];
+
+  // Events
+  #[cfg(feature = "events")]
+  let event_tx = {
+    let addr = comm.get_addr();
+    let (event_tx, event_thrds) = events::spawn_event_handlers(
+      node_config.ws.unwrap_or_default(),
+      node_config.ui,
+      addr,
+    );
+    threads.extend(event_thrds);
+    event_tx
+  };
+
+  // Mining
+  let (miner_comm, miner_thrds) =
+    spawn_miner(node_config.mining, event_tx.clone());
+  threads.extend(miner_thrds.into_iter());
+
+  // Node state object
+  let (node_query_sender, node) = Node::new(
+    node_config.data_path,
+    node_config.network_id,
+    initial_peers,
+    comm,
+    miner_comm,
+    #[cfg(feature = "events")]
+    event_tx,
+  );
+
+  // Spawns the API thread
+  if let Some(api_config) = api_config {
+    let api_thread = std::thread::spawn(move || {
+      crate::api::server::http_api_loop(node_query_sender, api_config);
+    });
+    threads.push(api_thread);
+  }
+
+  // Spawns the node thread
+  let node_thread = std::thread::spawn(move || {
+    node.main();
+  });
+  threads.insert(0, node_thread);
+
+  // Joins all threads
+  for thread in threads {
+    thread.join().unwrap();
+  }
 }
